Move encryption funcs to secrets.go
