+++ store/providers/xorm/models.go
@@ -1,9 +1,6 @@
 package xorm
 
 import (
-	"context"
-
-	"gocloud.dev/secrets"
 	"gomodules.xyz/secrets/xkms"
 	"k8s.io/klog/klogr"
 )
@@ -22,28 +19,4 @@ func init() {
 		new(SSHKey),
 		new(xkms.SecretKey),
 	)
-}
-
-func encryptData(secretID string, data []byte) ([]byte, error) {
-	ctx := context.Background()
-	urlstr := xkms.Scheme + "://" + secretID
-	keeper, err := secrets.OpenKeeper(ctx, urlstr)
-	if err != nil {
-		return nil, err
-	}
-	defer keeper.Close()
-
-	return keeper.Encrypt(ctx, data)
-}
-
-func decryptData(secretID string, cipher []byte) ([]byte, error) {
-	ctx := context.Background()
-	urlstr := xkms.Scheme + "://" + secretID
-	keeper, err := secrets.OpenKeeper(ctx, urlstr)
-	if err != nil {
-		return nil, err
-	}
-	defer keeper.Close()
-
-	return keeper.Decrypt(ctx, cipher)
 }
